BZ-1193085 critical IO error on client disconnect
When a Netty OIO thread servicing a client performs work on the journal
(e.g. creating a new page file, etc.) there is a small window where if
the connection is closed (either via administrative intervention or
because of ping failure, etc.) the Netty thread can be interrupted and
HornetQ will interpret this as a critical journal failure and shut
itself down.